camelcasify model filenames
