+++ packages/web/src/models/Todo.ts
@@ -2,9 +2,9 @@ import type { UUID } from 'crypto'
 import { v4 as uuidv4 } from 'uuid'
 import { Collection } from 'lokijs'
 
-import { Interval } from './interval'
-import { Tally } from './tally'
-import { Comment } from './comment'
+import { Interval } from './Interval'
+import { Tally } from './Tally'
+import { Comment } from './Comment'
 
 // TODO: Prefer not to do this; what alternatives exist?
 import { useIntervalsStore } from '@/stores/intervals'
