+++ packages/web/src/models/__tests__/Todo.spec.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest'
-import { Todo } from '../todo'
+import { Todo } from '../Todo'
 
 describe.todo('Todo', () => {
 })
